LayoutTests/fast/css-grid-layout/grid-columns-rows-get-set-expected.txt

Test that setting and getting grid-columns and grid-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting |display| set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-rows') is 'none'
PASS getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-columns') is '10px'
PASS getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-rows') is '15px'
PASS getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-columns') is '53%'
PASS getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-rows') is '27%'
PASS getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-columns') is 'auto'
PASS getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-rows') is 'auto'
PASS getComputedStyle(gridWithEMElement, '').getPropertyValue('-webkit-grid-columns') is '100px'
PASS getComputedStyle(gridWithEMElement, '').getPropertyValue('-webkit-grid-rows') is '150px'

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is 'none'

Test getting and setting display through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '18px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '66px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '55%'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '40%'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is 'auto'

Test setting grid-columns and grid-rows back to 'none' through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is '18px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is '66px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-rows') is 'none'
PASS successfullyParsed is true

TEST COMPLETE
